backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.2
rapidfuzz>=3.1.1
pandas>=2.0.0
numpy>=1.24.0
python-multipart>=0.0.6
openai-whisper>=20230918
pymongo>=4.6.0
